simplify / avoid code duplication + skip tests that have no way to work at this poing
