+++ bsi_zoo/tests/test_cross_val.py
@@ -5,34 +5,37 @@ import pytest
 
 from bsi_zoo.data_generator import get_data
 from bsi_zoo.estimators import (
-    # iterative_L1,
-    # iterative_L2,
-    # iterative_sqrt,
-    # iterative_L1_typeII,
-    # iterative_L2_typeII,
-    gamma_map,
-    TemporalCVSolver,
+    iterative_L1,
+    iterative_L2,
+    iterative_sqrt,
+    iterative_L1_typeII,
+    iterative_L2_typeII,
+    gamma_map
+)
+from bsi_zoo.cross_val import (
+    SpatialCVSolver,
+    TemporalCVSolver
 )
 
 
-@pytest.mark.parametrize("n_times", [100])
+@pytest.mark.parametrize("cv_type", ["temporal", "spatial"])
 @pytest.mark.parametrize("orientation_type", ["fixed", "free"])
 @pytest.mark.parametrize("nnz", [3])
 @pytest.mark.parametrize(
     "estimator,rtol,atol,cov_type,extra_params",
     [
-        # (iterative_L1, 1e-1, 5e-1, "diag", {}),
-        # (iterative_L2, 1e-1, 5e-1, "diag", {}),
-        # (iterative_sqrt, 1e-1, 5e-1, "diag", {}),
-        # (iterative_L1_typeII, 1e-1, 5e-1, "full", {}),
-        # (iterative_L2_typeII, 1e-1, 5e-1, "full", {}),
+        (iterative_L1, 1e-1, 5e-1, "diag", {}),
+        (iterative_L2, 1e-1, 5e-1, "diag", {}),
+        (iterative_sqrt, 1e-1, 5e-1, "diag", {}),
+        (iterative_L1_typeII, 1e-1, 5e-1, "full", {}),
+        (iterative_L2_typeII, 1e-1, 5e-1, "full", {}),
         (gamma_map, 1e-1, 5e-1, "full", {"update_mode": 1}),
-        # (gamma_map, 1e-1, 5e-1, "full", {"update_mode": 2}),
-        # (gamma_map, 1e-1, 5e-1, "full", {"update_mode": 3}),
+        (gamma_map, 1e-1, 5e-1, "full", {"update_mode": 2}),
+        (gamma_map, 1e-1, 5e-1, "full", {"update_mode": 3}),
     ],
 )
-def test_run_temporal_cv(
-    n_times,
+def test_cv(
+    cv_type,
     estimator,
     rtol,
     atol,
@@ -41,6 +44,16 @@ def test_run_temporal_cv(
     orientation_type,
     extra_params,
 ):
+    assert cv_type in ["temporal", "spatial"]
+
+    if estimator in [iterative_L1_typeII, iterative_L2_typeII] and cov_type == "full":
+        pytest.skip("iterative L1 type 2 breaks with full covariance")
+
+    if cv_type == "spatial":
+        n_times = 5
+    else:
+        n_times = 100
+
     n_orient = 1 if orientation_type == "fixed" else 3
 
     y, L, x, cov, noise = get_data(
@@ -64,15 +77,28 @@ def test_run_temporal_cv(
     else:
         alphas = [0.1, 0.01, 0.001]
 
-    solver = TemporalCVSolver(
-        estimator,
-        alphas=alphas,
-        cov_type=cov_type,
-        cov=cov,
-        n_orient=n_orient,
-        cv=2,
-        extra_params=extra_params,
-    ).fit(L=L, y=y)
+    if cv_type == "spatial":
+        solver = SpatialCVSolver(
+            estimator,
+            alphas=alphas,
+            cov_type=cov_type,
+            cov=cov,
+            n_orient=n_orient,
+            cv=3,
+            extra_params=extra_params,
+        )
+    else:
+        solver = TemporalCVSolver(
+            estimator,
+            alphas=alphas,
+            cov_type=cov_type,
+            cov=cov,
+            n_orient=n_orient,
+            cv=2,
+            extra_params=extra_params
+        )
+
+    solver.fit(L=L, y=y)
     x_hat = solver.predict(y)
 
     # check that the estimated noise level is correct
